add checkscan for data, and no use bulk insert in scanning internal

+++ media/java/android/media/MediaScanner.java
@@ -1287,14 +1287,27 @@ public class MediaScanner
         }
     }
 
+    //delete some uri in databases ,whese _data are null
+    //these uri are produced in eject and process.media died
+    private void checkProvider(String volumeName) throws RemoteException {
+        if (!volumeName.equals("internal")) {
+            String where = FileColumns.DATA + "=?";
+            String[] whereArgs = new String[] {""};
+            mMediaProvider.delete(mFilesUri, where, whereArgs);
+        }
+    }
+    
     public void scanDirectories(String[] directories, String volumeName) {
         try {
             long start = System.currentTimeMillis();
             initialize(volumeName);
+            
+            checkProvider(volumeName);
+            
             prescan(null, true);
             long prescan = System.currentTimeMillis();
 
-            if (ENABLE_BULK_INSERTS) {
+            if (ENABLE_BULK_INSERTS  && !volumeName.equals("internal")) {
                 // create MediaInserter for bulk inserts
                 mMediaInserter = new MediaInserter(mMediaProvider, 500);
             }
@@ -1303,7 +1316,7 @@ public class MediaScanner
                 processDirectory(directories[i], mClient);
             }
 
-            if (ENABLE_BULK_INSERTS) {
+            if (ENABLE_BULK_INSERTS && !volumeName.equals("internal")) {
                 // flush remaining inserts
                 mMediaInserter.flushAll();
                 mMediaInserter = null;
@@ -1313,7 +1326,7 @@ public class MediaScanner
             postscan(directories);
             long end = System.currentTimeMillis();
 
-            if (false) {
+            if (true) {
                 Log.d(TAG, " prescan time: " + (prescan - start) + "ms\n");
                 Log.d(TAG, "    scan time: " + (scan - prescan) + "ms\n");
                 Log.d(TAG, "postscan time: " + (end - scan) + "ms\n");
@@ -1327,6 +1340,10 @@ public class MediaScanner
             Log.e(TAG, "UnsupportedOperationException in MediaScanner.scan()", e);
         } catch (RemoteException e) {
             Log.e(TAG, "RemoteException in MediaScanner.scan()", e);
+        }finally {
+            if(mMediaProvider != null){
+                mMediaProvider = null;
+            }
         }
     }
 
@@ -1507,6 +1524,7 @@ public class MediaScanner
                 return new FileEntry(rowId, path, lastModified, format);
             }
         } catch (RemoteException e) {
+        } catch (SQLException e){
         } finally {
             if (c != null) {
                 c.close();
